# Conflicts:
#	.gitignore
#	src/main/java/com/board/model/BoardCategoryVO.java
#	src/main/java/com/board/model/BoardVO.java
#	src/main/java/com/board/model/PagingVO.java
#	src/main/java/com/common/CommonUtil.java
#	src/main/java/com/multi/webiting/AdminController.java
#	src/main/java/com/multi/webiting/BoardController.java
#	src/main/java/com/multi/webiting/UserController.java
#	src/main/java/com/product/model/CartVO.java
#	src/main/java/com/product/model/CategoryVO.java
#	src/main/java/com/product/model/ProductVO.java
#	src/main/java/com/user/mapper/AdminMapper.java
#	src/main/java/com/user/model/NotUserException.java
#	src/main/java/com/user/model/UserVO.java
#	src/main/java/com/user/service/AdminService.java
#	src/main/java/com/user/service/AdminServiceImpl.java
#	src/main/java/com/user/service/KakaoService.java
#	src/main/java/com/user/service/UserService.java
#	src/main/java/com/user/service/UserServiceImpl.java
#	src/main/resources/config/mybatis/mapper/AdminMapper.xml
#	src/main/resources/config/mybatis/mapper/BoardMapper.xml
#	src/main/resources/config/mybatis/mybatis-config.xml
#	src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
#	src/main/webapp/WEB-INF/views/admin/prodList.jsp
#	src/main/webapp/WEB-INF/views/board/boardEdit.jsp
#	src/main/webapp/WEB-INF/views/board/boardList.jsp
#	src/main/webapp/WEB-INF/views/board/boardView.jsp
#	src/main/webapp/WEB-INF/views/board/boardWrite.jsp
#	src/main/webapp/WEB-INF/views/foot.jsp
#	src/main/webapp/WEB-INF/views/header.jsp
#	src/main/webapp/WEB-INF/views/home.jsp
#	src/main/webapp/WEB-INF/views/index.jsp
#	src/main/webapp/WEB-INF/views/member/editform.jsp
#	src/main/webapp/WEB-INF/views/member/userlist.jsp
#	src/main/webapp/WEB-INF/views/msg.jsp
#	src/main/webapp/WEB-INF/views/top.jsp
#	src/main/webapp/resources/js/scripts.js
#	target/classes/com/common/CommonUtil.class
#	target/classes/com/common/interceptor/AdminCheckInterceptor.class
#	target/classes/com/common/interceptor/LoginCheckInterceptor.class
#	target/classes/com/multi/webiting/AdminController.class
#	target/classes/com/multi/webiting/UserController.class
#	target/classes/com/product/service/PAdminService.class
#	target/classes/com/product/service/PAdminServiceImpl.class
#	target/classes/com/product/service/ShopService.class
#	target/classes/com/product/service/ShopServiceImpl.class
#	target/classes/com/user/mapper/AdminMapper.class
#	target/classes/com/user/mapper/UserMapper.class
#	target/classes/com/user/model/NotUserException.class
#	target/classes/com/user/model/UserVO.class
#	target/classes/com/user/service/AdminService.class
#	target/classes/com/user/service/AdminServiceImpl.class
#	target/classes/com/user/service/KakaoService.class
#	target/classes/com/user/service/UserService.class
#	target/classes/com/user/service/UserServiceImpl.class
#	target/classes/config/mybatis/mapper/AdminMapper.xml
#	target/classes/config/mybatis/mapper/BoardMapper.xml
#	target/classes/config/mybatis/mybatis-config.xml
#	target/m2e-wtp/web-resources/META-INF/maven/com.multi/webiting/pom.properties

+++ src/main/java/com/multi/webiting/AdminController.java
@@ -1,8 +1,6 @@
 package com.multi.webiting;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import javax.inject.Inject;
 import javax.servlet.http.HttpServletRequest;
@@ -18,8 +16,8 @@ import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import com.common.CommonUtil;
 import com.board.model.PagingVO;
+import com.common.CommonUtil;
 import com.user.model.UserVO;
 import com.user.service.AdminService;
 
@@ -35,42 +33,77 @@ public class AdminController {
 	
 	CommonUtil common = new CommonUtil();
 	
-	@GetMapping("/userList")
-	public String userList(Model m, @ModelAttribute("page") PagingVO page,
-			HttpServletRequest req, @RequestHeader("User-Agent") String userAgent) {
-		HttpSession ses=req.getSession();
-		log.info("1. page===="+page);
-		int totalCount=this.aService.getTotalCount();
-		page.setTotalCount(totalCount);
-		page.setPagingBlock(10);
-		page.init(ses);
-		log.info("2. page===="+page);
-		List<UserVO> userArr=aService.listUser(null);
-		String loc="admin/list";
-		//String pageNavi=page.getPageNavi(loc, userAgent);
+	@GetMapping("/userList_old")
+	public String userList(Model m) {
+		int totalCount=aService.getTotalCount();
+		//List<UserVO> arr=aService.listUser();
 		
-		//m.addAttribute("pageNavi", pageNavi);
-		m.addAttribute("paging", page);
-		m.addAttribute("userArr", userArr);
+		//m.addAttribute("userArr",arr);		
+		m.addAttribute("totalCount",totalCount);
+		
 		return "member/userlist";
 	}
 	
+	//PagingVO에 getPageNavi()추가 (paging java로 처리)ㅡ 검색 기능 추가
+	@GetMapping("/userList")
+	public String userListPaging(Model m, @ModelAttribute("page") PagingVO page,
+			HttpServletRequest req, @RequestHeader("User-Agent") String userAgent) {
+		String myctx=req.getContextPath();
+		
+		HttpSession ses=req.getSession();
+		
+		log.info("1. page===="+page);
+		int totalCount=this.aService.getTotalCount(page);
+		page.setTotalCount(totalCount);
+		page.setPagingBlock(5);
+		page.init(ses);
+		
+		log.info("2. page===="+page);
+		List<UserVO> userArr=this.aService.selectUserAllPaging(page);
+		String loc="admin/userList";
+		String pageNavi=page.getPageNavi(myctx, loc, userAgent);
+		
+		m.addAttribute("pageNavi", pageNavi);
+		m.addAttribute("paging", page);
+		m.addAttribute("userArr", userArr);
+		return "member/userlist";
+		
+	}
+	
+	
 	@PostMapping("/userDel")
-	public String userDelete(@RequestParam(defaultValue="0") int idx) {
+	public String userDelete(Model m, @RequestParam(defaultValue="0") int idx) {
 		log.info("idx==="+idx);
+		
 		if(idx==0) {
 			return "redirect:userList";
 		}
+		
+		UserVO vo=this.aService.selectUserByIdx(idx);
+		if(vo==null) {
+			return common.addMsgBack(m, "해당 회원은 존재하지 않습니다.");
+		}
+		
 		int n=aService.deleteUser(idx);
 		
-		return "redirect:userList";
+		String str=(n>0)?"삭제 되었습니다.":"회원 삭제 실패";
+		String loc=(n>0)?"userList":"javascript:history.back()";
+		
+		return common.addMsgLoc(m, str, loc);
 	}
 	
 	@PostMapping("/userEdit")
-	public String userEdit(Model m, @RequestParam int idx) {
+	public String userEdit(Model m, @RequestParam(defaultValue="0") int idx) {
 		log.info("idx==="+idx);
 		
+		if(idx==0) {
+			return "redirect:userList";
+		}
+		
 		UserVO vo=this.aService.selectUserByIdx(idx);
+		if(vo==null) {
+			return common.addMsgBack(m, "해당 회원은 존재하지 않습니다.");
+		}
 		
 		m.addAttribute("user", vo);
